Refacto: rename all client functions
