+++ clients/geo/departements.ts
@@ -8,7 +8,7 @@ interface IGeoDepartementResponse {
   code: string;
 }
 
-const searchDepartements = async (slug: string): Promise<any> => {
+const clientGeoDepartements = async (slug: string): Promise<any> => {
   const response = await httpGet(routes.geo.departement + slug, {
     timeout: constants.timeout.L,
   });
@@ -27,4 +27,4 @@ const mapToDomainObject = (
   });
 };
 
-export { searchDepartements };
+export { clientGeoDepartements };
