+++ clients/open-data-soft/bodacc.ts
@@ -45,7 +45,7 @@ interface IBodaccB extends IBodaccCoreRecord {
   radiationaurcs?: string; // "{\"radiationPM\": \"O\"}"
 }
 
-const fetchAnnoncesBodacc = async (siren: Siren): Promise<IAnnoncesBodacc> => {
+const clientBodacc = async (siren: Siren): Promise<IAnnoncesBodacc> => {
   const searchUrl = `${routes.bodacc.ods.search}&q=registre%3A${siren}&sort=dateparution&facet=publicationavis&facet=publicationavis_facette&facet=typeavis&facet=typeavis_lib&facet=familleavis&facet=familleavis_lib&facet=numerodepartement&facet=departement_nom_officiel`;
   const metadataUrl = routes.bodacc.ods.metadata;
   const response = await odsClient(searchUrl, metadataUrl);
@@ -150,4 +150,4 @@ const extractDetails = (annonce: IBodaccRecords): string => {
   }
 };
 
-export default fetchAnnoncesBodacc;
+export default clientBodacc;
