+++ clients/ping-api-clients.ts
@@ -1,13 +1,13 @@
 import { verifyIdRna } from '../utils/helpers/id-rna';
 import { verifySiren } from '../utils/helpers/siren-and-siret';
-import { fetchAssociation } from './rna';
+import { clientRNA } from './rna';
 import { fetchRNCSImmatriculationNoCache } from './rncs';
-import { fetchRnmImmatriculation } from './rnm';
-import { getUniteLegaleInseeNoCache } from './sirene-insee/siren';
-import getResults from './sirene-ouverte/recherche';
-import getUniteLegaleSireneOuverte from './sirene-ouverte/siren';
-import fetchConventionCollectives from './siret-2-idcc';
-import { validateTVANumber } from './tva';
+import { clientRNM } from './rnm';
+import { clientUniteLegaleInseeNoCache } from './sirene-insee/siren';
+import clientSearchSireneOuverte from './sirene-ouverte/recherche';
+import clientUniteLegaleSireneOuverte from './sirene-ouverte/siren';
+import clientSiret2Idcc from './siret-2-idcc';
+import { clientTVA } from './tva';
 
 export class APISlugNotFound extends Error {
   constructor(public status: number, public message: string) {
@@ -22,19 +22,19 @@ const ping = async (slug: string | string[]) => {
       // fetch IRM and disable cache
       return await fetchRNCSImmatriculationNoCache(verifySiren('880878145'));
     case 'api-rnm':
-      return await fetchRnmImmatriculation(verifySiren('824024350'));
+      return await clientRNM(verifySiren('824024350'));
     case 'api-conventions-collectives':
-      return await fetchConventionCollectives(['54205118000066']);
+      return await clientSiret2Idcc(['54205118000066']);
     case 'api-sirene-insee':
-      return await getUniteLegaleInseeNoCache(verifySiren('880878145'));
+      return await clientUniteLegaleInseeNoCache(verifySiren('880878145'));
     case 'api-sirene-donnees-ouvertes':
-      return await getUniteLegaleSireneOuverte(verifySiren('880878145'));
+      return await clientUniteLegaleSireneOuverte(verifySiren('880878145'));
     case 'api-rna':
-      return await fetchAssociation(verifyIdRna('W551000280'), useCache);
+      return await clientRNA(verifyIdRna('W551000280'), useCache);
     case 'api-tva':
-      return await validateTVANumber(verifySiren('880878145'), useCache);
+      return await clientTVA(verifySiren('880878145'), useCache);
     case 'api-recherche':
-      return await getResults('test', 1, undefined);
+      return await clientSearchSireneOuverte('test', 1, undefined);
     default:
       throw new APISlugNotFound(404, `API ping ${slug} not found`);
   }
