+++ clients/rncs/rncs-proxy-client.ts
@@ -9,7 +9,7 @@ import httpClient, { httpGet } from '../../utils/network';
  * @param useCache by default, cache is deactivated
  * @returns
  */
-const APIRncsProxyGet = async (
+const clientRNCSProxy = async (
   url: string,
   options?: AxiosRequestConfig,
   useCache = false // by default we dont cache response
@@ -26,22 +26,4 @@ const APIRncsProxyGet = async (
     useCache
   );
 
-/**
- * POST for RNCS Proxy api - no cache
- *
- * @param url
- * @param options
- * @returns
- */
-const APIRncsProxyPost = async (url: string, options: AxiosRequestConfig) =>
-  await httpClient({
-    url,
-    ...options,
-    headers: {
-      'X-APIkey': process.env.PROXY_API_KEY || '',
-      ...options?.headers,
-    },
-    method: 'POST',
-  });
-
-export { APIRncsProxyGet, APIRncsProxyPost };
+export { clientRNCSProxy };
