+++ clients/siret-2-idcc/index.ts
@@ -33,7 +33,7 @@ interface IConventionObject {
   siret?: string;
 }
 
-const fetchConventionCollectives = async (
+const clientSiret2Idcc = async (
   sirets: string[]
 ): Promise<IConventionCollective[]> => {
   const batches = generateBatches(sirets);
@@ -71,4 +71,4 @@ const mapToDomainObject = (
   }, []);
 };
 
-export default fetchConventionCollectives;
+export default clientSiret2Idcc;
