+++ models/tva.ts
@@ -1,4 +1,4 @@
-import { validateTVANumber } from '../clients/tva';
+import { clientTVA } from '../clients/tva';
 import { Siren, verifySiren } from '../utils/helpers/siren-and-siret';
 import logErrorInSentry from '../utils/sentry';
 
@@ -31,11 +31,11 @@ export const tvaIntracommunautaire = async (
   const tvaNumberFromSiren = tvaNumber(siren);
 
   try {
-    return await validateTVANumber(tvaNumberFromSiren);
+    return await clientTVA(tvaNumberFromSiren);
   } catch {
-    // retry once
+    // retry once as VIES randomely reset connection
     try {
-      return await validateTVANumber(tvaNumberFromSiren);
+      return await clientTVA(tvaNumberFromSiren);
     } catch (eFallback: any) {
       logErrorInSentry('Error in API TVA', {
         details: eFallback.toString(),
