+++ pages/api/geo/[slug].ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { searchCommunes } from '../../../clients/geo/communes';
-import { searchDepartements } from '../../../clients/geo/departements';
+import { clientGeoCommunes } from '../../../clients/geo/communes';
+import { clientGeoDepartements } from '../../../clients/geo/departements';
 import logErrorInSentry from '../../../utils/sentry';
 import { withAPM } from '../../../utils/sentry/apm';
 
@@ -11,8 +11,8 @@ const geo = async (
   try {
     const term = slug as string;
     const [communes, departements] = await Promise.all([
-      searchCommunes(term),
-      searchDepartements(term),
+      clientGeoCommunes(term),
+      clientGeoDepartements(term),
     ]);
 
     const results = [...departements.slice(0, 5), ...communes.slice(0, 10)];
